+++ CHANGELOG.md
@@ -8,6 +8,9 @@ have notable changes.
 
 Changes since v2.19
 
+## Fixes
+- Show the selected but disabled resources in Change resources autocomplete (#2702, #2693)
+
 ## v2.19 "Nahkahousuntie" 2021-06-28
 
 ### Fixes